requirements.txt
# Core dependencies
python-telegram-bot==21.1
requests
apscheduler>=3.10.0

# Webhook server
flask==3.1.2

# AI-powered moderation (optional - commented out due to dependency conflicts)
# httpx==0.25.2
# groq==0.4.2

# Optional: Pattern matching (falls back to regex if unavailable)
# pyahocorasick==2.1.0  # Optional C-extension: requires build tools on Windows

# OpenAI API (optional - commented out)
# openai

# Local AI (optional; for toxicity detection)
# transformers
# torch
